requirements.txt
aniso8601==8.0.0
attrs==20.3.0
click==7.1.2
coverage==5.3
flake8==3.8.4
Flask==1.1.2
Flask-Cors==3.0.9
Flask-RESTful==0.3.8
Flask-Testing==0.8.0
geographiclib==1.50
geopy==2.0.0
iniconfig==1.1.1
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mccabe==0.6.1
numpy==1.19.4
opencv-python==4.4.0.46
packaging==20.4
Pillow==8.0.1
pluggy==0.13.1
price-parser==0.3.3
py==1.9.0
pycodestyle==2.6.0
pyflakes==2.2.0
pyparsing==2.4.7
pytesseract==0.3.6
pytest==6.1.2
pytz==2020.4
six==1.15.0
toml==0.10.2
urllib3==1.26.2
Werkzeug==1.0.1
